app-py/requirements.txt
aerich==0.9.0
aiomysql==0.2.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncclick==8.1.8.0
atlastk==0.13.4
cffi==1.17.1
chardet==4.0.0
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
dictdiffer==0.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fonttools==4.58.0
h11==0.16.0
idna==3.10
iso8601==2.1.0
kiwisolver==1.4.8
matplotlib==3.9.2
packaging==25.0
passlib==1.7.4
pillow==11.2.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyMySQL==1.1.1
pyparsing==3.2.3
pypika-tortoise==0.6.0
python-dateutil==2.9.0.post0
python-jose==3.4.0
pytz==2025.2
rsa==4.9.1
scipy==1.14.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tortoise==0.1.1
tortoise-orm==0.25.0
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
xlrd==2.0.1
pandas==3.0.2
python-dotenv==1.2.2
python-multipart==0.0.22
aiohttp==3.13.4
geopandas==1.1.3
# torch 建议优先用 conda 安装 CPU 版；若你用 pip 安装遇到 DLL/依赖问题，按我前面给的 conda 命令重装 torch
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
